perf: Wrap the slot request handler in useCallback
Based on a preliminary benchmark using the React Dev Tools,
it seems to reduce the amount of re-renders
The actual performance numbers show minor improvement,
but the test case was a combo component which is not that
template heavy and React is plenty fast on re-renders anyway.

A better benchmark scenario would be a heavily templated grid.

Once the React compiler is officially stable, this may become
obsolete, since presumably it would optimize this case, as well
as others in applications using the package.